+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Blackcube Core - CHANGELOG 
 ==========================
 
+ * Add: Cached files management
+ * Add: Assets file management
+ * Add: sitemap.xml generation
+ * Add: robots.txt generation
+ * Upd: Update dependencies
+
 Release 3.1.1 October, 27th 2023 <pgaultier@redcat.io>
 ----------------------------------------------------